corrections

+++ 0x03-debugging/2-largest_number.c
@@ -1,8 +1,7 @@
 #include "main.h"
 
 /**
- * largest_number - returns the largest of 3
- * integers
+ * largest_number - returns the largest of 3 numbers
  * @a: first integer
  * @b: second integer
  * @c: third integer
@@ -10,7 +9,7 @@
  * Return: largest number
  */
 
-long largest_number(int a, int b, int c)
+long largest_number(long a, long b, long c)
 {
 	int largest_N;
