+++ 0x03-debugging/main.h
@@ -4,7 +4,7 @@
 #include <stdio.h>
 
 void positive_or_negative(int i);
-long largest_number(int a, int b, int c);
+long largest_number(long a, long b, long c);
 void print_remaining_days(int day, int month, int year);
 int convert_day(int month, int day);
 #endif /* MAIN_H */
